Complete cutover
